move tab group constants to templates service
